GRIFFON: Get rid of config.h/cpp
