+++ engines/griffon/module.mk
@@ -2,7 +2,6 @@ MODULE := engines/griffon
 
 MODULE_OBJS := \
 	combat.o \
-	config.o \
 	console.o \
 	cutscenes.o \
 	detection.o \
